+++ packages/eslint-plugin/src/rules/no-unnecessary-type-arguments.ts
@@ -17,7 +17,7 @@ export default util.createRule<[], MessageIds>({
   meta: {
     docs: {
       description:
-        'Warns if an explicitly specified type argument is the default for that type parameter',
+        'Enforces that type arguments will not be used if not required',
       category: 'Best Practices',
       recommended: false,
       requiresTypeChecking: true,
@@ -109,6 +109,10 @@ function getTypeParametersFromType(
   }
 
   const sym = getAliasedSymbol(symAtLocation, checker);
+
+  if (!sym.declarations) {
+    return undefined;
+  }
 
   return findFirstResult(sym.declarations, decl =>
     tsutils.isClassLikeDeclaration(decl) ||