+++ packages/eslint-plugin/src/rules/no-untyped-public-signature.ts
@@ -9,11 +9,10 @@ type MessageIds = 'noReturnType' | 'untypedParameter';
 type Options = [{ ignoredMethods: string[] }];
 
 export default util.createRule<Options, MessageIds>({
-  name: 'no-unused-public-signature',
+  name: 'no-untyped-public-signature',
   meta: {
     docs: {
-      description:
-        'Requires that all public method arguments and return type will be explicitly typed',
+      description: 'Disallow untyped public methods',
       category: 'Best Practices',
       recommended: false,
     },