+++ packages/eslint-plugin/tests/rules/prefer-readonly.test.ts
@@ -13,9 +13,9 @@ const ruleTester = new RuleTester({
 
 ruleTester.run('prefer-readonly', rule, {
   valid: [
-    `function ignore() { }`,
-    `const ignore = function () { }`,
-    `const ignore = () => { }`,
+    'function ignore() { }',
+    'const ignore = function () { }',
+    'const ignore = () => { }',
     `const container = { member: true };
       container.member;`,
     `const container = { member: 1 };
@@ -30,7 +30,7 @@ ruleTester.run('prefer-readonly', rule, {
       --container.member;`,
     `const container = { member: 1 };
       container.member--;`,
-    `class TestEmpty { }`,
+    'class TestEmpty { }',
     `class TestReadonlyStatic {
       private static readonly correctlyReadonlyStatic = 7;
     }`,